+++ src/collective/contact/mailaction/adapters.py
@@ -13,6 +13,11 @@ class IRecipientProvider(Interface):
 
 @implementer(IRecipientProvider)
 class RecipientProvider(object):
+    """return the name and email of a person
+    as defined in rfc2822 "3.4. Address Specification"
+
+    if the person has no email set - none is returned
+    """
 
     adapts(IPerson)
 
@@ -21,6 +26,9 @@ class RecipientProvider(object):
 
     def __call__(self):
         cd = IContactDetails(self.context)
+        if not cd.email:
+            return None
+
         if self.context.firstname:
             return formataddr((
                 "%s %s" % (self.context.firstname, self.context.lastname),